更新資料庫

+++ EnjoyEat/EnjoyEat/Models/DTO/ProductDTO.cs
@@ -12,7 +12,7 @@ namespace EnjoyEat.Models.DTO
         public string? MealImg { get; set; }
         public string? ProductName { get; set; } = null!;
         public short UnitPrice { get; set; }
-        public byte Costs { get; set; }
+        public short Costs { get; set; }
         public short? Stock { get; set; } = 0!;
         public string? Description { get; set; }
         public string? Recipe { get; set; }
